Fix JAX scatter type incompatibility warnings
Closes #24

Eliminates all 22+ FutureWarnings about scatter type mismatches that will
become errors in future JAX releases.

**Changes (11 fixes across 7 files)**:

**jaxnav environments (3 fixes)**:
- jaxnav_ued_utils.py: Cast to goal_map_mask.dtype for scatter ops (2)
- jaxnav_env.py: Cast to lidar_hits.dtype for boolean scatter (1)
- maps/grid_map.py: Dtype-aware casting for map masks (4)

**smax environment (2 fixes)**:
- smax_env.py: Use jnp.float32(1) and jnp.uint8(1) for typed arrays (2)

**storm environments (2 fixes)**:
- storm.py: Use False instead of 0 for boolean conflicts array (2)
- storm_env.py: Use jnp.bool_() for interacted_agents matrix (2)

**tests (4 fixes)**:
- test_smax.py: Use True/False for boolean done arrays (4)

**Error Fixed**:
```
FutureWarning: scatter inputs have incompatible types:
cannot safely cast value from dtype=int32 to dtype=bool
```

**Test Results**:
✅ All affected tests pass (jaxnav, smax, storm)
✅ 0 remaining scatter warnings
✅ 100% backward compatible

**Verification**:
Run with warnings enabled to confirm zero scatter warnings:
python -W default::FutureWarning -m pytest tests/storm/ tests/jaxnav/ tests/smax/

+++ jaxmarl/environments/smax/smax_env.py
@@ -786,7 +786,7 @@ class SMAX(MultiAgentEnv):
                 max_segment_angle = (2 * math.pi) * (
                     (j + 1) / self.num_sections
                 ) - math.pi
-                self_mask = jnp.zeros((self.num_agents,)).at[i].set(1)
+                self_mask = jnp.zeros((self.num_agents,)).at[i].set(jnp.float32(1))
                 in_range_mask = (
                     (angle > min_segment_angle)
                     & (angle < max_segment_angle)
@@ -913,7 +913,7 @@ class SMAX(MultiAgentEnv):
             is_alive = state.unit_alive[i]
             mask = jnp.zeros((num_actions,), dtype=jnp.uint8)
             # always can take the stop action
-            mask = mask.at[self.num_movement_actions - 1].set(1)
+            mask = mask.at[self.num_movement_actions - 1].set(jnp.uint8(1))
             mask = mask.at[: self.num_movement_actions - 1].set(
                 jax.lax.select(
                     is_alive,
